update C3P0 for more compatiblity.
The PoolBackedDataSource  class within this payload is not nesscessary, and  it does not declare a static serialVersionUID, that may get different serialVersionUID from various versions of C3P0. But the PoolBackedDataSource  class has a hard-coded serialVersionUID,  so removing PoolBackedDataSource can make this payload more compatible with other versions of C3P0.